output example
- Move `dnsjit.output.null` to `example.output.null`
- `input-example`:
  - Clarify module name for logging
  - Fix test `require()`
- Disabled tests using `test_pcap_read.lua`
