+++ examples/modules/README.md
@@ -1,6 +1,6 @@
 # TODO
 + move zero to input
 - create counter filter
-- move null to output
++ move null to output
 - pure Lua lib example
 - move test_pcap_read/test_throughput to example dnsjit programs
