+++ examples/modules/input-example/test.lua
@@ -1,9 +1,12 @@
 local ffi = require("ffi")
-local zero = require("dnsjit.input.zero").new()
+local zero = require("example.input.zero").new()
 
 local producer, context = zero:produce()
 local object = producer(context)
 
 if ffi.istype("struct core_object*", object) then
     print("loading and usage successful")
+    os.exit(0)
 end
+
+os.exit(1)
