+++ examples/modules/output-example/Makefile.am
@@ -1,0 +1,11 @@
+MAINTAINERCLEANFILES = $(srcdir)/Makefile.in \
+  $(srcdir)/configure
+
+SUBDIRS = src
+
+dist_doc_DATA = README.md
+
+EXTRA_DIST = test.lua
+
+test:
+	dnsjit $(srcdir)/test.lua
