+++ examples/modules/output-example/src/output/null.c
@@ -1,32 +1,20 @@
-/*
- * Copyright (c) 2018-2021, OARC, Inc.
- * All rights reserved.
- *
- * This file is part of dnsjit.
- *
- * dnsjit is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * dnsjit is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with dnsjit.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 #include "config.h"
 
-#include "output/null.h"
-#include "core/assert.h"
-#include "core/object/pcap.h"
+#include <dnsjit/core/assert.h>
+#include <dnsjit/core/log.h>
+#include <dnsjit/core/receiver.h>
+#include <dnsjit/core/producer.h>
 
-static core_log_t    _log      = LOG_T_INIT("output.null");
+typedef struct output_null {
+    core_log_t      _log;
+    core_producer_t prod;
+    void*           ctx;
+    size_t          pkts;
+} output_null_t;
+
+static core_log_t    _log      = LOG_T_INIT("example.output.null");
 static output_null_t _defaults = {
-    LOG_T_INIT_OBJ("output.null"),
+    LOG_T_INIT_OBJ("example.output.null"),
     0, 0, 0
 };
