--- src/output/null.h
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2018-2021, OARC, Inc.
- * All rights reserved.
- *
- * This file is part of dnsjit.
- *
- * dnsjit is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * dnsjit is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with dnsjit.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include <dnsjit/core/log.h>
-#include <dnsjit/core/receiver.h>
-#include <dnsjit/core/producer.h>
-
-#ifndef __dnsjit_output_null_h
-#define __dnsjit_output_null_h
-
-#include <stddef.h>
-#include <stdint.h>
-
-#include <dnsjit/output/null.hh>
-
-#endif
