--- src/output/null.hh
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2018-2021, OARC, Inc.
- * All rights reserved.
- *
- * This file is part of dnsjit.
- *
- * dnsjit is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * dnsjit is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with dnsjit.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-//lua:require("dnsjit.core.log")
-//lua:require("dnsjit.core.receiver_h")
-//lua:require("dnsjit.core.producer_h")
-
-typedef struct output_null {
-    core_log_t      _log;
-    core_producer_t prod;
-    void*           ctx;
-    size_t          pkts;
-} output_null_t;
-
-core_log_t* output_null_log();
-void        output_null_init(output_null_t* self);
-void        output_null_destroy(output_null_t* self);
-void        output_null_run(output_null_t* self, int64_t num);
-
-core_receiver_t output_null_receiver();
